requirements.txt
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
h5py>=3.0.0
pandas>=1.3.0
joblib>=1.1.0
pillow>=8.0.0
opencv-python>=4.5.0
torch>=1.9.0
torchvision>=0.10.0
clip-by-openai>=1.0
nibabel>=3.2.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0
annoy>=1.17.0
jupyter>=1.0.0
ipykernel>=6.0.0
alipy>=1.0.0
statsmodels>=0.12.0
umap-learn>=0.5.0
hdbscan>=0.8.0
google-generativeai>=0.3.0
pathlib2>=2.3.0
concurrent-futures>=3.1.0